Add cs_open/cs_close use SafeHandle(s)
 - Renamed DisassembleErrorCode to CapstoneErrorCode
 - Moved ExceptionThrower class content into Helpers class
 - Modified SafeCapstoneHandle to get rid of Microsoft.Win32
dependency and made IntPtr.Zero the only invalid handle
value.
 - Modified Capstone exported cs_close signature for interop
and made it internal only. This implied using a custom
marshaler to handle the ref modifier of the pHandle parameter.
 - Modified Capstone exported cs_open signature for interop
and made it internal only. The returned handle is now decorated
with an out modified instead of a ref one.
- Modified Capstone exported cs_option signature for interop
and made it internal only.